~ renamed project **test.scratch.string_slices.c** => **test.unit.string_slices.c** ; ~ warnings
