+++ test/scratch/CMakeLists.txt
@@ -4,7 +4,6 @@ add_subdirectory(test.scratch.basics2)
 add_subdirectory(test.scratch.fail_all.1)
 add_subdirectory(test.scratch.integer_comparisons.c)
 add_subdirectory(test.scratch.setup.fail)
-add_subdirectory(test.scratch.string_slices.c)
 add_subdirectory(test.scratch.temp_directory)
 add_subdirectory(test.scratch.temp_file)
 add_subdirectory(test.scratch.verbosity_levels)
