+++ test/unit/CMakeLists.txt
@@ -5,3 +5,4 @@ if(shwild_FOUND)
 endif(shwild_FOUND)
 
 add_subdirectory(test.unit.string_comparisons)
+add_subdirectory(test.unit.string_slices.c)
