+++ test/unit/test.unit.string_slices.c/CMakeLists.txt
@@ -1,2 +1,2 @@
 # SIS:AUTO_GENERATED: Remove this line if you edit the file, otherwise it will be overwritten
-define_automated_test_program(test.scratch.string_slices.c test.scratch.string_slices.c.c)
+define_automated_test_program(test.unit.string_slices.c test.unit.string_slices.c.c)
