+++ test/unit/test.unit.string_slices.c/vc10/test.unit.string_slices.c.vcxproj
@@ -159,7 +159,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\test.scratch.string_slices.c.c" />
+    <ClCompile Include="..\test.unit.string_slices.c.c" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
