+++ test/unit/test.unit.string_slices.c/vc6/test.unit.string_slices.c.dsp
@@ -1,24 +1,24 @@
-# Microsoft Developer Studio Project File - Name="test.scratch.string_slices.c" - Package Owner=<4>
+# Microsoft Developer Studio Project File - Name="test.unit.string_slices.c" - Package Owner=<4>
 # Microsoft Developer Studio Generated Build File, Format Version 6.00
 # ** DO NOT EDIT **
 
 # TARGTYPE "Win32 (x86) Application" 0x0101
 
-CFG=test.scratch.string_slices.c - Win32 Debug
+CFG=test.unit.string_slices.c - Win32 Debug
 !MESSAGE This is not a valid makefile. To build this project using NMAKE,
 !MESSAGE use the Export Makefile command and run
 !MESSAGE
-!MESSAGE NMAKE /f "test.scratch.string_slices.c.mak".
+!MESSAGE NMAKE /f "test.unit.string_slices.c.mak".
 !MESSAGE
 !MESSAGE You can specify a configuration when running NMAKE
 !MESSAGE by defining the macro CFG on the command line. For example:
 !MESSAGE
-!MESSAGE NMAKE /f "test.scratch.string_slices.c.mak" CFG="test.scratch.string_slices.c - Win32 Debug"
+!MESSAGE NMAKE /f "test.unit.string_slices.c.mak" CFG="test.unit.string_slices.c - Win32 Debug"
 !MESSAGE
 !MESSAGE Possible choices for configuration are:
 !MESSAGE
-!MESSAGE "test.scratch.string_slices.c - Win32 Release" (based on "Win32 (x86) Application")
-!MESSAGE "test.scratch.string_slices.c - Win32 Debug" (based on "Win32 (x86) Application")
+!MESSAGE "test.unit.string_slices.c - Win32 Release" (based on "Win32 (x86) Application")
+!MESSAGE "test.unit.string_slices.c - Win32 Debug" (based on "Win32 (x86) Application")
 !MESSAGE
 
 # Begin Project
@@ -29,7 +29,7 @@ CPP=cl.exe
 MTL=midl.exe
 RSC=rc.exe
 
-!IF  "$(CFG)" == "test.scratch.string_slices.c - Win32 Release"
+!IF  "$(CFG)" == "test.unit.string_slices.c - Win32 Release"
 
 # PROP BASE Use_MFC 0
 # PROP BASE Use_Debug_Libraries 0
@@ -56,7 +56,7 @@ LINK32=link.exe
 # ADD LINK32 kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /nologo /version:1.0 /subsystem:console /map /machine:I386
 # SUBTRACT LINK32 /pdb:none
 
-!ELSEIF  "$(CFG)" == "test.scratch.string_slices.c - Win32 Debug"
+!ELSEIF  "$(CFG)" == "test.unit.string_slices.c - Win32 Debug"
 
 # PROP BASE Use_MFC 0
 # PROP BASE Use_Debug_Libraries 1
@@ -80,21 +80,21 @@ BSC32=bscmake.exe
 # ADD BSC32 /nologo
 LINK32=link.exe
 # ADD BASE LINK32 /nologo /subsystem:windows /debug /machine:I386 /pdbtype:sept
-# ADD LINK32 kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /nologo /version:1.0 /subsystem:console /map /debug /machine:I386 /out:"Debug/test.scratch.string_slices.c.debug.exe" /pdbtype:sept
+# ADD LINK32 kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /nologo /version:1.0 /subsystem:console /map /debug /machine:I386 /out:"Debug/test.unit.string_slices.c.debug.exe" /pdbtype:sept
 # SUBTRACT LINK32 /pdb:none
 
 !ENDIF
 
 # Begin Target
 
-# Name "test.scratch.string_slices.c - Win32 Release"
-# Name "test.scratch.string_slices.c - Win32 Debug"
+# Name "test.unit.string_slices.c - Win32 Release"
+# Name "test.unit.string_slices.c - Win32 Debug"
 # Begin Group "Source Files"
 
 # PROP Default_Filter "cpp;c;cxx;rc;def;r;odl;idl;hpj;bat"
 # Begin Source File
 
-SOURCE=..\test.scratch.string_slices.c.c
+SOURCE=..\test.unit.string_slices.c.c
 # End Source File
 # End Group
 # Begin Group "Header Files"
