+++ xtests.vc10.sln
@@ -40,7 +40,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "test.scratch.setup.fail", "
 		{B1C8AE78-41F1-4176-AE0C-DEE491424EA8} = {B1C8AE78-41F1-4176-AE0C-DEE491424EA8}
 	EndProjectSection
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "test.scratch.string_slices.c", "test\scratch\test.scratch.string_slices.c\vc10\test.scratch.string_slices.c.vcxproj", "{DD4ECD42-CA57-4963-B39C-E7F7FB25291D}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "test.unit.string_slices.c", "test\scratch\test.unit.string_slices.c\vc10\test.unit.string_slices.c.vcxproj", "{DD4ECD42-CA57-4963-B39C-E7F7FB25291D}"
 	ProjectSection(ProjectDependencies) = postProject
 		{B1C8AE78-41F1-4176-AE0C-DEE491424EA8} = {B1C8AE78-41F1-4176-AE0C-DEE491424EA8}
 	EndProjectSection
